@@ -1,9 +1,1 @@
-/// <reference types: string
-/// <reference types: string
-/// <reference types: string
-
-export {};
-
-declare global {
-	var jQuery: JQueryStatic
-}
+/// <reference types="@sveltejs/kit" />
