@@ -1,7 +1,7 @@
 <script>
   import { Meta, Template, Story } from '@storybook/addon-svelte-csf'
-  import Grid, { Column, Row } from '../../src/collections/grid'
-  import Icon from '../../src/elements/icon'
+  import Grid, { Column, Row } from '../../src/lib/collections/grid'
+  import Icon from '../../src/lib/elements/icon'
 </script>
 
 <Meta
