@@ -1,6 +1,6 @@
 <script>
     import { Meta, Story } from '@storybook/addon-svelte-csf'
-    import Search from '../../src/modules/search'
+    import Search from '../../src/lib/modules/search'
   
     /**
      * @type {Search}
